Implements image display in lists
Adds support for displaying images within the list component.

This change introduces a new `isImageColumn` property to the `TableColumn` interface, allowing developers to designate columns as image display columns. It also provides options for specifying an `imageSrc` function to retrieve the image source from the data and a `defaultImage` to use if no source is available.

Also includes minor fixes:

- Updates the person component to use a default image.
- Moves the book import template to a more appropriate directory.